Oclbridge implementation started

+++ Java/Bridge/src/main/java/flinkOcl/OclContext.java
@@ -2,6 +2,13 @@ package flinkOcl;
 
 import configuration.*;
 import flinkOcl.buildEngine.BuildEngine;
+import flinkOcl.buildEngine.CppLibraryInfo;
+import oclBridge.OclBridge;
+
+import java.io.File;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Objects;
 
 public class OclContext
 {
@@ -10,6 +17,9 @@ public class OclContext
 	private IUserFunctionsRepository mFunctionRepository;
 	
 	private String mApplicationDirectory;
+	private CppLibraryInfo mCppLibraryInfo;
+	
+	private OclBridge mOclBridgeContext;
 	
 	public OclContext(ISettingsRepository pSettingsRepository,
 					  ITupleDefinitionsRepository pTupleDefinitionsRepository,
@@ -30,22 +40,43 @@ public class OclContext
 	public void open()
 	{
 		createAndBuildAndLoadKernels();
+		
+		mOclBridgeContext.initialize();
 	}
 	
 	public void close()
 	{
-		deleteLocalFiles();
+		mOclBridgeContext.dispose();
+		
+		if(mSettingsRepository.getContextOptions().hasToRemoveTempFoldersOnClose())
+			deleteLocalFiles();
 	}
 
 	private void createAndBuildAndLoadKernels()
 	{
-		new BuildEngine(mSettingsRepository)
-				.generateKernels(mTupleDefinitionsRepository, mFunctionRepository.getUserFunctions())
-				.loadCppLibrary();
+		BuildEngine vBuildEngine = new BuildEngine(mSettingsRepository)
+				.generateKernels(mTupleDefinitionsRepository, mFunctionRepository.getUserFunctions());
+		
+		vBuildEngine.loadCppLibrary();
+		mCppLibraryInfo = vBuildEngine.getCppLibraryInfo();
 	}
 	
 	private void deleteLocalFiles()
 	{
-		//TODO usare getApplicationDirectory per eliminare i file aggiunti
+		Path vKernelsFolder = Paths.get(mCppLibraryInfo.getKernelsFolder());
+		
+		File vToFile = vKernelsFolder.toFile();
+		
+		if (vToFile.isFile())
+			return;
+		
+		boolean vAllFilesDeleted = true;
+		for (File vFile : Objects.requireNonNull(vToFile.listFiles()))
+		{
+			vAllFilesDeleted &= vFile.delete();
+		}
+		
+		if (vAllFilesDeleted)
+			vToFile.delete();
 	}
 }
